requirements.txt
#Click==7.0
uvicorn==0.20.0
fastapi>=0.20.0,<0.113.0
sqlalchemy>=2.0.0
pydantic>=2.7.0,<3.0.0
psycopg2>=1.0.0
httpx==0.27.0
pymongo==4.7.1
python-jose==3.3.0
python-decouple==3.1
passlib==1.7.4
bcrypt==4.1.3
#h11==0.9.0
#httptools==0.0.13
#psycopg2-binary==2.8.4
#psycopg2==2.9.9
#pydantic==2.7
#SQLAlchemy==1.3.12
#starlette==0.12.9
#typing==3.7.4.1
#uvicorn==0.11.1
#uvloop==0.14.0
#websockets==8.1